% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomfuncs.R
\name{add2}
\alias{add2}
\title{Function for summing two objects}
\usage{
add2(x, y)
}
\arguments{
\item{x}{An R object}

\item{y}{An R object}
}
\description{
Adds two R objects of the same dimensions
}
\examples{
ex1 <- add2(2,3);
ex2 <- add2(c(1,1,2),c(1,2,2));
}
