% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomfuncs.R
\name{square}
\alias{square}
\title{Square function}
\usage{
square(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Takes the square of the R object
}
\examples{
ex1 <- square(2);
ex2 <- square(c(2,3));
}
